feat(cli)!: remove sync command
