+++ docs/clawhub.md
@@ -14,7 +14,7 @@ sidebarTitle: "ClawHub"
 ClawHub is the public registry for OpenClaw skills and plugins.
 
 - Use native `openclaw` commands to search, install, and update skills and to install plugins from ClawHub.
-- Use the separate `clawhub` CLI for registry auth, publishing, sync, and delete/undelete workflows.
+- Use the separate `clawhub` CLI for registry auth, publishing, and delete/undelete workflows.
 
 Site: [clawhub.ai](https://clawhub.ai)
 
@@ -37,7 +37,7 @@ openclaw plugins update --all
 ```
 
 Install the ClawHub CLI when you want registry-authenticated workflows such as
-publish, sync, or delete/undelete:
+publish or delete/undelete:
 
 ```bash
 npm i -g clawhub
@@ -85,7 +85,6 @@ clawhub package explore --family code-plugin
 clawhub package inspect episodic-claw
 clawhub package publish your-org/your-plugin --dry-run
 clawhub package publish your-org/your-plugin
-clawhub sync --all
 ```
 
 The CLI also has skill install/update commands for direct registry workflows:
