+++ docs/publishing.md
@@ -33,120 +33,6 @@ clawhub skill publish ./my-skill \
 Use `--owner <handle>` when publishing to an org owner. Omit it to publish as
 the authenticated user.
 
-For catalog repos, use `sync` to scan folders containing `SKILL.md` and publish
-new or changed skills:
-
-```bash
-clawhub sync --dry-run --owner <owner>
-clawhub sync --all --owner <owner>
-```
-
-Use `--dry-run` first to see the plan without uploading.
-
-### GitHub Actions for Skills
-
-If you want to run skill publishing from CI, call ClawHub's reusable
-[`skill-publish.yml` workflow](https://github.com/openclaw/clawhub/blob/main/.github/workflows/skill-publish.yml)
-from a small workflow in your repo.
-
-The example below is shaped for a catalog repo: operators choose whether to
-preview the full catalog, publish one skill folder, or publish the whole
-catalog.
-
-```yaml
-name: Publish Skills to ClawHub
-
-on:
-  workflow_dispatch:
-    inputs:
-      mode:
-        description: What to run.
-        type: choice
-        required: true
-        default: dry-run
-        options:
-          - dry-run
-          - publish-single
-          - publish-catalog
-      skill_path:
-        description: Skill folder for publish-single, for example skills/<slug>.
-        type: string
-        required: false
-        default: ""
-
-permissions:
-  contents: read
-  id-token: write
-
-jobs:
-  validate-single:
-    if: github.event_name == 'workflow_dispatch' && inputs.mode == 'publish-single'
-    runs-on: ubuntu-latest
-    steps:
-      - name: Validate single-skill input
-        env:
-          SKILL_PATH: ${{ inputs.skill_path }}
-        run: |
-          set -euo pipefail
-          if [[ -z "${SKILL_PATH}" ]]; then
-            echo "::error::skill_path is required when mode is publish-single."
-            exit 1
-          fi
-          case "${SKILL_PATH}" in
-            skills/*) ;;
-            *)
-              echo "::error::skill_path must point under skills/, for example skills/<slug>."
-              exit 1
-              ;;
-          esac
-
-  dry-run:
-    if: github.event_name == 'workflow_dispatch' && inputs.mode == 'dry-run'
-    uses: openclaw/clawhub/.github/workflows/skill-publish.yml@main
-    with:
-      owner: <owner>
-      dry_run: true
-    secrets:
-      clawhub_token: ${{ secrets.CLAWHUB_TOKEN }}
-
-  publish-single:
-    if: github.event_name == 'workflow_dispatch' && inputs.mode == 'publish-single'
-    needs: validate-single
-    uses: openclaw/clawhub/.github/workflows/skill-publish.yml@main
-    with:
-      owner: <owner>
-      skill_path: ${{ inputs.skill_path }}
-      dry_run: false
-    secrets:
-      clawhub_token: ${{ secrets.CLAWHUB_TOKEN }}
-
-  publish-catalog:
-    if: github.event_name == 'workflow_dispatch' && inputs.mode == 'publish-catalog'
-    uses: openclaw/clawhub/.github/workflows/skill-publish.yml@main
-    with:
-      owner: <owner>
-      dry_run: false
-    secrets:
-      clawhub_token: ${{ secrets.CLAWHUB_TOKEN }}
-```
-
-Replace `<owner>` with your ClawHub owner handle. The called workflow defaults to
-scanning `skills/`; pass `skill_path` only when you want to process one folder.
-
-Before running a real publish, sign in as a ClawHub user that can publish to the
-selected owner, then store the current CLI token as a `CLAWHUB_TOKEN` repository
-secret:
-
-```bash
-clawhub login --label "Skills GitHub Actions"
-gh secret set CLAWHUB_TOKEN \
-  --repo OWNER/REPO \
-  --body "$(clawhub token)"
-```
-
-Start with `dry-run`, then publish one skill with `publish-single`, and only then
-use `publish-catalog` for the full catalog.
-
 ## Plugins
 
 Plugins use npm-style package names. Scoped package names include the owner in
