+++ packages/clawhub/src/cli/clawdbotConfig.ts
@@ -1,5 +1,5 @@
 import { readFile } from "node:fs/promises";
-import { basename, join, resolve } from "node:path";
+import { join, resolve } from "node:path";
 import JSON5 from "json5";
 import { resolveHome } from "../homedir.js";
 
@@ -9,59 +9,11 @@ type ClawdbotConfig = {
     defaults?: { workspace?: string };
     list?: Array<{
       id?: string;
-      name?: string;
       workspace?: string;
       default?: boolean;
     }>;
   };
-  routing?: {
-    agents?: Record<
-      string,
-      {
-        name?: string;
-        workspace?: string;
-      }
-    >;
-  };
-  skills?: {
-    load?: {
-      extraDirs?: string[];
-    };
-  };
 };
-
-type ClawdbotSkillRoots = {
-  roots: string[];
-  labels: Record<string, string>;
-};
-
-export async function resolveClawdbotSkillRoots(): Promise<ClawdbotSkillRoots> {
-  const roots: string[] = [];
-  const labels: Record<string, string> = {};
-
-  const clawdbotStateDir = resolveClawdbotStateDir();
-  const sharedSkills = resolveUserPath(join(clawdbotStateDir, "skills"));
-  pushRoot(roots, labels, sharedSkills, "Shared skills");
-
-  const openclawStateDir = resolveOpenclawStateDir();
-  const openclawShared = resolveUserPath(join(openclawStateDir, "skills"));
-  pushRoot(roots, labels, openclawShared, "OpenClaw: Shared skills");
-
-  const [clawdbotConfig, openclawConfig] = await Promise.all([
-    readClawdbotConfig(),
-    readOpenclawConfig(),
-  ]);
-  if (!clawdbotConfig && !openclawConfig) return { roots, labels };
-
-  if (clawdbotConfig) {
-    addConfigRoots(clawdbotConfig, roots, labels);
-  }
-  if (openclawConfig) {
-    addConfigRoots(openclawConfig, roots, labels, "OpenClaw");
-  }
-
-  return { roots, labels };
-}
 
 export async function resolveClawdbotDefaultWorkspace(): Promise<string | null> {
   const config = await readClawdbotConfig();
@@ -143,62 +95,4 @@ async function readConfigFile(path: string): Promise<ClawdbotConfig | null> {
   } catch {
     return null;
   }
-}
-
-function addConfigRoots(
-  config: ClawdbotConfig,
-  roots: string[],
-  labels: Record<string, string>,
-  labelPrefix?: string,
-) {
-  const prefix = labelPrefix ? `${labelPrefix}: ` : "";
-
-  const mainWorkspace = resolveUserPath(
-    config.agents?.defaults?.workspace ?? config.agent?.workspace ?? "",
-  );
-  if (mainWorkspace) {
-    pushRoot(roots, labels, join(mainWorkspace, "skills"), `${prefix}Agent: main`);
-  }
-
-  const listedAgents = config.agents?.list ?? [];
-  for (const entry of listedAgents) {
-    const workspace = resolveUserPath(entry?.workspace ?? "");
-    if (!workspace) continue;
-    const name = entry?.name?.trim() || entry?.id?.trim() || "agent";
-    pushRoot(roots, labels, join(workspace, "skills"), `${prefix}Agent: ${name}`);
-  }
-
-  const agents = config.routing?.agents ?? {};
-  for (const [agentId, entry] of Object.entries(agents)) {
-    const workspace = resolveUserPath(entry?.workspace ?? "");
-    if (!workspace) continue;
-    const name = entry?.name?.trim() || agentId;
-    pushRoot(roots, labels, join(workspace, "skills"), `${prefix}Agent: ${name}`);
-  }
-
-  const extraDirs = config.skills?.load?.extraDirs ?? [];
-  for (const dir of extraDirs) {
-    const resolved = resolveUserPath(dir);
-    if (!resolved) continue;
-    const label = `${prefix}Extra: ${basename(resolved) || resolved}`;
-    pushRoot(roots, labels, resolved, label);
-  }
-}
-
-function pushRoot(roots: string[], labels: Record<string, string>, root: string, label?: string) {
-  const resolved = resolveUserPath(root);
-  if (!resolved) return;
-  if (!roots.includes(resolved)) roots.push(resolved);
-  if (!label) return;
-  const existing = labels[resolved];
-  if (!existing) {
-    labels[resolved] = label;
-    return;
-  }
-  const parts = existing
-    .split(", ")
-    .map((part) => part.trim())
-    .filter(Boolean);
-  if (parts.includes(label)) return;
-  labels[resolved] = `${existing}, ${label}`;
 }
